+++ src/main/kotlin/restApi/GroupChatImage.kt
@@ -1,6 +1,6 @@
 package com.neelkamath.omniChatBackend.restApi
 
-import com.neelkamath.omniChatBackend.db.PicType
+import com.neelkamath.omniChatBackend.db.ImageType
 import com.neelkamath.omniChatBackend.db.tables.GroupChatUsers
 import com.neelkamath.omniChatBackend.db.tables.GroupChats
 import com.neelkamath.omniChatBackend.userId
@@ -10,33 +10,33 @@ import io.ktor.http.*
 import io.ktor.response.*
 import io.ktor.routing.*
 
-fun routeGroupChatPic(routing: Routing): Unit = with(routing) {
-    route("group-chat-pic") {
-        getGroupChatPic(this)
-        authenticate { patchGroupChatPic(this) }
+fun routeGroupChatImage(routing: Routing): Unit = with(routing) {
+    route("group-chat-image") {
+        getGroupChatImage(this)
+        authenticate { patchGroupChatImage(this) }
     }
 }
 
-private fun getGroupChatPic(route: Route): Unit = with(route) {
+private fun getGroupChatImage(route: Route): Unit = with(route) {
     get {
         val chatId = call.parameters["chat-id"]!!.toInt()
-        val type = PicType.valueOf(call.parameters["pic-type"]!!)
+        val type = ImageType.valueOf(call.parameters["image-type"]!!)
         if (GroupChats.isExisting(chatId)) {
-            val pic = GroupChats.readPic(chatId, type)
-            if (pic == null) call.respond(HttpStatusCode.NoContent) else call.respondBytes(pic)
+            val image = GroupChats.readImage(chatId, type)
+            if (image == null) call.respond(HttpStatusCode.NoContent) else call.respondBytes(image)
         } else call.respond(HttpStatusCode.BadRequest)
     }
 }
 
-private fun patchGroupChatPic(route: Route): Unit = with(route) {
+private fun patchGroupChatImage(route: Route): Unit = with(route) {
     patch {
         val chatId = call.parameters["chat-id"]!!.toInt()
-        val pic = readMultipartPic()
+        val image = readMultipartImage()
         when {
-            pic == null -> call.respond(HttpStatusCode.BadRequest)
+            image == null -> call.respond(HttpStatusCode.BadRequest)
             !GroupChatUsers.isAdmin(call.userId!!, chatId) -> call.respond(HttpStatusCode.Unauthorized)
             else -> {
-                GroupChats.updatePic(chatId, pic)
+                GroupChats.updateImage(chatId, image)
                 call.respond(HttpStatusCode.NoContent)
             }
         }