+++ src/test/kotlin/db/MediaTest.kt
@@ -1,24 +1,34 @@
 package com.neelkamath.omniChatBackend.db
 
 import com.neelkamath.omniChatBackend.readBytes
-import com.neelkamath.omniChatBackend.readPic
+import com.neelkamath.omniChatBackend.readImage
 import org.junit.Test
 import org.junit.jupiter.api.Nested
 import kotlin.test.assertEquals
 import kotlin.test.assertFailsWith
 import kotlin.test.assertNotEquals
 
-class PicTest {
+class ImageTest {
     @Nested
     inner class Init {
         @Test
         fun `Passing an excessively large original image must cause an exception to be thrown`() {
-            assertFailsWith<IllegalArgumentException> { Pic.build("png", ByteArray(Pic.ORIGINAL_MAX_BYTES + 1)) }
+            assertFailsWith<IllegalArgumentException> {
+                ProcessedImage.build(
+                    "png",
+                    ByteArray(ProcessedImage.ORIGINAL_MAX_BYTES + 1)
+                )
+            }
         }
 
         @Test
         fun `Passing an excessively large thumbnail must cause an exception to be thrown`() {
-            assertFailsWith<IllegalArgumentException> { Pic(ByteArray(1), ByteArray(Pic.THUMBNAIL_MAX_BYTES + 1)) }
+            assertFailsWith<IllegalArgumentException> {
+                ProcessedImage(
+                    ByteArray(1),
+                    ByteArray(ProcessedImage.THUMBNAIL_MAX_BYTES + 1)
+                )
+            }
         }
     }
 
@@ -27,7 +37,7 @@ class PicTest {
     inner class Type_Companion_build {
         @Test
         fun `Using a valid capitalized file extension mustn't fail`() {
-            Pic.Type.build("PNG")
+            ProcessedImage.Type.build("PNG")
         }
     }
 
@@ -36,12 +46,12 @@ class PicTest {
     inner class Companion_build {
         @Test
         fun `Passing a supported extension must work`() {
-            Pic.build("jpg", readPic("76px×57px.jpg").original)
+            ProcessedImage.build("jpg", readImage("76px×57px.jpg").original)
         }
 
         @Test
         fun `Passing an unsupported extension mustn't work`() {
-            assertFailsWith<IllegalArgumentException> { Pic.build("webp", readBytes("76px×57px.webp")) }
+            assertFailsWith<IllegalArgumentException> { ProcessedImage.build("webp", readBytes("76px×57px.webp")) }
         }
     }
 }
@@ -50,21 +60,21 @@ class MediaTest {
     @Nested
     inner class CreateThumbnail {
         @Test
-        fun `A thumbnail for a pic larger than 100px by 100px must be at most 100px by 100px`(): Unit =
-            with(readPic("1008px×756px.jpg")) { assertNotEquals(original, thumbnail) }
+        fun `A thumbnail for an image larger than 100px by 100px must be at most 100px by 100px`(): Unit =
+            with(readImage("1008px×756px.jpg")) { assertNotEquals(original, thumbnail) }
 
         @Test
-        fun `A thumbnail for a pic which is smaller than 100px by 100px must be the same as the pic`(): Unit =
-            with(readPic("76px×57px.jpg")) { assertEquals(thumbnail, original) }
+        fun `A thumbnail for an image which is smaller than 100px by 100px must be the same as the image`(): Unit =
+            with(readImage("76px×57px.jpg")) { assertEquals(thumbnail, original) }
 
         @Test
         fun `A thumbnail must be built for a JPEG image`() {
-            readPic("76px×57px.jpg")
+            readImage("76px×57px.jpg")
         }
 
         @Test
         fun `A thumbnail must be built for a PNG image`() {
-            readPic("76px×57px.png")
+            readImage("76px×57px.png")
         }
     }
 }